Make the power menu configurable [1/2]
This allows to toggle all the options that we've added and some AOSP ones

- Reboot
- Advanced reboot (depends on reboot menu)
- Screenshot

Went ahead and commented out the other AOSP options in case a user wants to
get cute and toggle these on via a overlay. Also anymore than 5 options and
things start to look janky.

aosip edit: remove airplane option

Change-Id: I5d86cef181a8f12ad18e16ccc1e8c75879b3e569
Signed-off-by: Josh Fox(XlxFoXxlX) <joshfox87@gmail.com>